requirements.txt
fastapi
uvicorn
pydantic
python-multipart
azure-storage-file-datalake
python-dotenv
